requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
spotipy==2.25.1
urllib3==2.3.0
ytmusicapi==1.10.2
